+++ source/isaaclab_assets/isaaclab_assets/cart_pendulum/mdp/__init__.py
@@ -3,4 +3,7 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-from .rewards import *
+"""Các term MDP riêng của dự án con lắc đơn trên xe đẩy."""
+
+from .commands import *  # noqa: F403
+from .rewards import *  # noqa: F403